AWF #361: keep gc.py under line cap + fix worker entrypoint test wiring
Iteration 1 fixes two regressions surfaced by running the validation gate
against commit 910e53dc1:

- gc.py grew to 1536 lines, breaching the 1500-line first-party
  maintainability guardrail (test_first_party_code_files_stay_under_line_limit).
  Extract the auth-overlay unmount helpers (_unmount_candidate_auth_overlay,
  _auth_overlay_unmount_skips_target, _auth_unmount_skipped_outcome) into a
  focused gc_auth_overlay.py sibling (the established gc_*.py pattern) and
  re-import them, preserving the awf.service.gc.<name> surface. gc.py is back
  to 1454 lines; the unused subprocess import is dropped.
- test_worker_entrypoint_wires_control_worker_dependencies used a _ControlWorker
  stub that predated the auth_overlay_work_dir wiring; update the stub to accept
  the kwarg and assert it equals work_dir, mirroring tests/unit/service/test_worker.py.

Behavior unchanged; gc_auth_overlay.py is 100% covered by the existing gc suites.
